+++ CHANGELOG.md
@@ -9,8 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.1.12] - 2020-01-28
 
+### Added
+
+- Support for Python type stub files (`*.pyi`).
+
 ### Changed
 
+- Fix yamllint path issue.
 - Fix issue where black would not run due to "mach-o, wrong architecture" in dependencies.
 
 ## [1.1.11] - 2020-01-27